add all PostController function create add and modify posts admin views

+++ routes/web.php
@@ -61,8 +61,8 @@ Route::get('/shop', function () {
 /* POST ROUTE */
 Route::get('admin/posts', 'Admin\PostController@index')->name('admin.posts.index');
 Route::get('admin/posts/create', 'Admin\PostController@create')->name('admin.posts.create');
-Route::post('posts', 'Admin\PostController@store')->name('admin.posts.store');
-Route::get('posts/{post}', 'Admin\PostController@show')->name('admin.posts.show');
-Route::get('posts/{post}/edit', 'AdminnPostController@edit')->name('admin.posts.edit');
-Route::put('posts/{post}', 'Admin\PostController@update')->name('admin.posts.update');
-Route::delete('posts/{post}', 'Admin\PostController@destroy')->name('admin.posts.destroy');
+Route::post('admin/posts', 'Admin\PostController@store')->name('admin.posts.store');
+Route::get('admin/posts/{post}', 'Admin\PostController@show')->name('admin.posts.show');
+Route::get('admin/posts/{post}/edit', 'Admin\PostController@edit')->name('admin.posts.edit');
+Route::put('admin/posts/{post}', 'Admin\PostController@update')->name('admin.posts.update');
+Route::delete('admin/posts/{post}', 'Admin\PostController@destroy')->name('admin.posts.destroy');
